Update card - rfid

+++ src/modules/card-rfid/card-rfid.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, UnprocessableEntityException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException, UnprocessableEntityException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CardRfid } from './entities/card-rfid.entity';
 import { Repository } from 'typeorm';
@@ -36,7 +36,7 @@ export class CardRfidService {
     }
 
     async findAllPaginate(options: IPaginationOptions): Promise<Pagination<CardRfid>> {
-        return paginate<CardRfid>(this.cardRfidRepository, options);
+        return await paginate<CardRfid>(this.cardRfidRepository, options, {order: {id: 'DESC'}});
     }
 
      async create(cardRfidDto: CardPlayLoadRfidDto) {
@@ -49,27 +49,38 @@ export class CardRfidService {
         const cardRfid = new CardRfid();
         cardRfid.name = cardRfidDto.name;
         cardRfid.card_uid = cardRfidDto.card_uid;
-        cardRfid.is_active = cardRfidDto.is_active;
+        cardRfid.is_activate = cardRfidDto.is_activate;
         cardRfid.description = cardRfidDto.description;
         return this.cardRfidRepository.save(cardRfid);
     }
 
     async findOne(id: number) {
-        return await this.cardRfidRepository.findOne({where: {id}});
+        const cardRfid = await this.cardRfidRepository.findOne({where: {id}});
+        if(!cardRfid) {
+            throw new UnprocessableEntityException('Card RFID not found');
+        }
+        return cardRfid;
     }
 
     async update(id: number, cardPlayLoadRfidDto: CardPlayLoadRfidDto) {
+        this.validateCardPlayLoad(cardPlayLoadRfidDto);
         const cardRfid = await this.cardRfidRepository.findOne({where: {id}});
+        if(!cardRfid) {
+            throw new UnprocessableEntityException('Card RFID not found');
+        }
         cardRfid.name = cardPlayLoadRfidDto.name;
         cardRfid.card_uid = cardPlayLoadRfidDto.card_uid;
-        cardRfid.is_active = cardPlayLoadRfidDto.is_active;
+        cardRfid.is_activate = cardPlayLoadRfidDto.is_activate;
         cardRfid.description = cardPlayLoadRfidDto.description;
         return this.cardRfidRepository.save(cardRfid);
     }
 
     async remove(id: number) {
         const cardRfid = await this.cardRfidRepository.findOne({where: {id}});
-        return this.cardRfidRepository.remove(cardRfid);
+        if(!cardRfid) {
+            throw new NotFoundException('Card RFID not found');
+        }
+        return await this.cardRfidRepository.remove(cardRfid);
     }
 
     async findByCardUid(card_uid: string) {
@@ -77,7 +88,7 @@ export class CardRfidService {
     }
 
     async findByCardUidAndIsActive(card_uid: string) {
-        return await this.cardRfidRepository.findOne({where: {card_uid, is_active: true}});
+        return await this.cardRfidRepository.findOne({where: {card_uid, is_activate: true}});
     }
 
     validateCardPlayLoad(cardPlayLoadRfidDto: CardPlayLoadRfidDto):boolean {
@@ -90,7 +101,7 @@ export class CardRfidService {
             throw new UnprocessableEntityException('Name is required');
         }
 
-        if(cardPlayLoadRfidDto.is_active === undefined) {
+        if(cardPlayLoadRfidDto.is_activate === undefined) {
             throw new UnprocessableEntityException('Is Active is required');
         }
 
@@ -99,5 +110,17 @@ export class CardRfidService {
         }
         
         return true;
+    }
+
+    async verifyCard(card_uid: string): Promise<{cardRfid: CardRfid, status: boolean}>
+    {   
+        if(card_uid === undefined) {
+            throw new NotFoundException('Card UID is required');
+        }
+        const cardRfid = await this.cardRfidRepository.findOne({where: {card_uid, is_activate: true}});
+        if(!cardRfid) {
+            throw new UnauthorizedException('Card RFID not found');
+        }
+        return {cardRfid,status: true};
     }
 }
